Moved getAppPath def
