+++ src/osp/util/ExecutablePath.h
@@ -1,0 +1,4 @@
+#pragma once
+#include <string>
+
+std::string getAppPath();
